useEffect 부분을 hook 으로 분리
