+++ src/components/LineUp.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 
 import styled from 'styled-components';
 
@@ -6,7 +6,7 @@ import Header from './Header';
 import LineupItem from './LineupItem';
 
 import useCheckScreenWidth from '../hooks/useCheckScreenWidth';
-import useFetchLineup from '../hooks/useFetchLineup';
+import useSetLineUpList from '../hooks/useSetLineUpList';
 
 const Content = styled.div<{ demoImgList: string[] }>`
   display: flex;
@@ -55,13 +55,9 @@ export default function LineUp() {
   const [perView, setPerView] = useState(3);
   const [spaceBetween, setSpaceBetween] = useState(50);
   const [demoImgList, setDemoImgList] = useState<string[]>([]);
-  const lineupList = useFetchLineup();
 
   useCheckScreenWidth({ setPerView, setSpaceBetween });
-  useEffect(() => {
-    const imgList = lineupList.map((lineup) => lineup.img);
-    setDemoImgList(imgList);
-  }, [lineupList]);
+  useSetLineUpList({ setDemoImgList });
 
   return (
     <>
